Summery App
I saved in an Excel file from a sheet to a table with the name of the application and other relevant information.
• A sheet for all the applications deleted from the computer.
I created 2 sheets in the same Excel file,
1. One is called Summary in which there are the names of all the applications from all the ways and a count of the number of impressions (you can draw conclusions from this in my opinion)
2. The second sheet Registry_Summary which contains the names of the applications I found through the Registry and the number of instances it appeared in the other ways, you can see that 120 apps are shared between all the ways.
(I chose the Registry to be my "compass" for the reason that it gives the same information as the information that windows gives according to the amount of applications that are installed on the computer which is 131 and you can see that the gap is 11. Apparently they were not registered in wmi and win api).